refactor: 스크린 범위 컴포넌트화 #1 by @hobak12
